--- src/app/(page)/(main)/(auth)/layout.tsx
@@ -1,13 +1,0 @@
-import ProtectedRoute from "@/app/components/auth/ProtectedRoute";
-
-export default async function AuthLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <>
-      <ProtectedRoute>{children}</ProtectedRoute>
-    </>
-  );
-}
