+++ src/app/(page)/(main)/(protected)/friend/page.tsx
