+++ src/app/(page)/(main)/(protected)/layout.tsx
@@ -1,0 +1,13 @@
+import ProtectedRoute from "@/components/auth/ProtectedRoute";
+
+export default async function AuthLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <ProtectedRoute>{children}</ProtectedRoute>
+    </>
+  );
+}
