+++ src/app/(page)/(main)/home/page.tsx
@@ -1,5 +1,5 @@
 import { cardService } from "@/core/Service/cardService";
-import CardRead from "../(auth)/mypage/_components/CardRead";
+import CardRead from "../../../../components/(card)/CardRead";
 
 export default async function HomePage() {
   const [cardAllDatas] = await Promise.all([
@@ -9,7 +9,7 @@ export default async function HomePage() {
   return (
     <div className="max-w-md mx-auto rounded-lg mb-32">
       {cardAllDatas.map((cardAllData, index) => (
-        <CardRead key={index} cardAllData={cardAllData} />
+        <CardRead key={index} cardAllData={cardAllData} isHome={true} />
       ))}
     </div>
   );
