--- src/app/components/general/Container/ColumnMainContainer.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-type ContainerProps = {
-  children: React.ReactNode
-}
-
-export default function ColumnMainContainer({ children }: ContainerProps) {
-  return <div className="">{children}</div>
-}
