--- src/app/components/general/index.ts
@@ -1,29 +1,0 @@
-import Button from './Button'
-import LoadingSpinner from './LoadingSpinner'
-import Modal from './Modal'
-import Divider from './Divider'
-// import Layout from "./Layout";
-// import AdminLayout from "./Layout/AdminLayout";
-// import Footer from "./Footer";
-import Container from './Container'
-
-export {
-  Button,
-  // BorderButton,
-  // BackButton,
-  Container,
-  // TwoColumnMainContainer,
-  // Header,
-  // Input,
-  // MultiLineTextField,
-  // Select,
-  // SearchableSelect,
-  LoadingSpinner,
-  Modal,
-  // Tag,
-  // Layout,
-  // AdminLayout,
-  Divider,
-  // Footer,
-  // Pagination,
-}
