+++ src/components/(news)/RichContent.tsx
